Tested with hardware but needs work for display
Shows correct value 50% of the time, need to improve to only show values
if the correct value is read.
Also it is not possible to change back from the measurement mode into
voltage or amp mode, as they use the same interrupt function. Also needs a
change. (only have the timer on for a certain amount of time for example)